Develop

+++ app/controllers/posts_controller.rb
@@ -1,7 +1,20 @@
 class PostsController < ApplicationController
 
   def index
-    @posts = Post.all
+    @posts = Post.page(params[:page]).page(params[:page]).per(5).order(created_at: :desc)
+  end
+
+  def rank
+    @posts = Post.find(Favorite.group(:post_id).pluck(:post_id))
+    @posts = Kaminari.paginate_array(@posts).page(params[:page]).per(5)
+  end
+
+  def random
+    if ActiveRecord::Base.connection_config[:adapter] == 'sqlite3'  #ローカル。DBで判断
+    @posts = Post.order("RANDOM()").limit(5)
+    else
+    @posts = Post.order("RAND()").limit(5)
+    end
   end
 
   def new
@@ -14,8 +27,7 @@ class PostsController < ApplicationController
     if @post.save
       redirect_to post_path(@post)
     else
-      @posts = Post.all
-      redirect_to posts_path
+      render 'new'
     end
   end
 
@@ -26,17 +38,28 @@ class PostsController < ApplicationController
 
   def edit
     @post = Post.find(params[:id])
+    if @user != current_user
+      redirect_to post_path(@post.id)
+    end
   end
 
   def update
     @post = Post.find(params[:id])
-    @post.update(post_params)
-    redirect_to post_path(@post)
+    if @post.update(post_params)
+      redirect_to post_path(@post)
+    else
+      render :edit
+    end
   end
 
   def destroy
     post = Post.find(params[:id])
-    post.destroy
+    if post.user == current_user
+      post.destroy
+      flash[:notice] = "投稿が削除されました。"
+    else
+      flash[:alert] = "投稿の削除に失敗しました。"
+    end
     redirect_to posts_path
   end
